separated benchmarks project
