+++ neo.Benchmarks/README.md
@@ -1,0 +1,4 @@
+NEO Benchmarks
+====================
+
+This project provides benchmarking and heavier tests not covered by regular unit tests.
