+++ neo.Benchmarks/protocol.json
@@ -1,0 +1,5 @@
+{
+  "ProtocolConfiguration": {
+    "SecondsPerBlock": 2
+  }
+}
